EventHandler: Deprecate Call(args, no_remote, ts), add Call(args)
The ts parameter was only added to Call() for the Broker::auto_publish()
functionality and propagating the network timestamp. By now, auto_publish
is deprecated, so it'd be good to cleanup that signature.

To not break the existing API though  and prepare for the scenario that
getting an event's timestamp will be lazy (instead of passing it around
unconditionally), move to a model where Call() invokes event_mgr.CurrentEventTime()
if ts=0.0 and the timestamp is actually used. The result is the same
it is now, but it allows to defer (or even skip) computing the event time
if the event is not going to be auto-published (which will be the case
most of the time going forward).