Event: Replace current_ts with current_meta
Instead of storing the timestamp in the event and event mgr, store
a metadata vector and compute the timestamp lazily if needed.

Most users aren't actually interested in the timestamp metadata, so
fetching it lazily should be okay. Further, holding a metadata vector
allows to more easily add further metadata values to events.

The only gripe is the added allocations for the vector and TimeVal in
the Event constructor.