Event: No need to be Obj
It's not fully clear to me why Event needs to be an Obj. Not
inheriting from Obj avoids needing a vtable and allows to skip
the Obj constructor and destructor for a bit of a performance
improvement in microbenchmarks.

Open-code Ref() and Unref() to keep the HookQueueEvent() API
stable for plugins that took charge of the Event (and need to free/unref it).