analyzer-failed.log - small fixes, start adjusting tests
The main part of this commit are changes in tests. A lot of the tests
that previously relied on analyzer.log or dpd.log now use the new
analyzer-failed.log.

There are still a couple of special cases left to cover, but this should
be the bulk of the changes.