Analyzer_failed.log: small tweaks and final test updates
This contains some small tweaks to analyzer_failed.log:

when the removal of an analyzer is vetoed, the analyzer_failed event is
now not raised.

If an analyzer is no longer active when an analyzer violation is raised,
currently the analyzer_failed event is raised. This can, e.g., happen
when an analyzer error happens at the very end of the connection. This
makes the behavior more similar to what happened in the past, and also
intuitively seems to make sense.

A bug introduced in the failed service logging was fixed.

The remaining test failures were fixed.

The external test suites were updated and adjusted for the new style
logs.