Properly finish Spicy TCP child analyzer
We would previously not `Finish` a Spicy TCP analyzer. This manifested
in a Spicy child analyzer fed with `protocol_data_in` not receiving an
EOD, so if it was waiting for that it would simply fail to finish
parsing.

We now explicitly call `Finish` when receiving `EndOfData`. That
function is also called from `Done`.