dce-rpc: Rename named_pipe to sec_addr
This is in response to #3935. I'm really not sure we should do that, but
it was important enough for someone to open an issue.

IMO, Zeek distributors could work around that by extending DCE_RPC::Info
with a new sec_addr field and removing `&log` from the original one. We
could provide a policy script to do that, but that seems little useful
to make progress towards removing the old field. So outright removing
and providing an option to add it back is what is proposed here.

Needs updates of external baselines. Skipping on that for now.