Add cleaned-up version of zeek-aux
This is the zeek-aux repository with all of the git-related stuff removed,
including all of the pre-commit configuration files, git directories,
and submodules. It has a modified version of the CMake configuration
that uses Zeek's variables and paths, making zeek-aux effectively just a
subdirectory to Zeek. All of the source files now have Zeek's license
header in them as well.