xdp-shunt: Start making it actually do stuff
This is still pretty bad - the drop function doesn't actually do what it
should and there is a bunch of hardcoded stuff. The BPF map formats
aren't exactly the same so it doesn't end up doing much, but it does
update the map!

All in all, it's a start, but no where near good.