ZAM: Fix memory leak in Construct-Known-Record-With-Inits
Commit b5d17f7f49c32e2b87856da01b51cf512c52e0e9 mistakenly removed Unref()
for init_vals pre-generated, but then potentially replaced by values from a
record. Re-instantiate it and add a test that tickled the leak.

This also replaced some autos with the concrete types, even if they
are a mouthful. Mostly to make my poor man's grep-based workflow a bit
easier. Also, given not everything is in .cc files, using an IDE to
figure out types is semi successful (works for the generated code only).