testing: Update baselines, add endpoint tests
The dump events test perturbs because many of the events never saw
up-to-date originator and responder state. E.g. the history field of a
connection now usually is S in new_connection() of TCP connections, while
previously it'd be empty. Also num_pkts and num_bytes_ip on the endpoint
aren't zero anymore. The first packet definitely has been observed.

Looking at this with --word-diff makes it somewhat reasonable.