Enforce length for 802.3 frames. Accept truncated 802.3 frames.
Non-ethernet-2 frames provide the length in one of their fields.

This commit switches to using the provided length during parsing. This
means that trailing bytes will now be ignored, whereas previously they
would have been passed on to downstream protocol analyzers.

This change aligns our parsing of 802.3 packets to the parsing of other
packet-types, where do enforce length fields (e.g. tcp/udp).

Furthermore, this commit now accepts trucated 802.3 frames; before they
were discarded. This similarly aligns us with other protocol parsers,
where this already was done; this also already was done with 802.3
frames wrapped in VLANs.

Thirdly, this commit changes the weird value `truncated_vlan_frame` to
`truncated_VLAN_header`. The weird was not updated when the check was
changed from a truncated frame check to a truncated header check. This
aligns the weird with another weird earlier in the file, and also makes
the handling the same as for normal ethernet packets in Ethernet.cc.

Addresses GH-4838