SNMPv3: add parsing UserSecurityModel Security Parameters
SNMPv3 has several security models; the typical one is the user-based
security model (USM) standardized in RFC3414. This commit adds parsing
of the UsmSecurityParameters when a v3 packet with the user-based
security model is encountered.

This, among other things, contains the username of the request which
(kind of) takes the place of the community string of earlier SNMP
versions. This data is added to the already existing header record.

Additionally, the username is added to snmp.log, taking the part of the
community string for SNMPv3.