cluster: Add Cluster::Table::publish_element_new
This adds a bif that can be used as an &on_change handler for distributing
every new element in a table across a Zeek cluster.

The implementation is based on a Cluster::Table::elements_new_internal()
event which is sent from within the change handler.

We should probably consider batching with a timer to make this more
efficient at the expense of slight delay.