analyzer: Introduce analyzer::for_each() helper
Allows to iterate through an analyzer tree recursively, calling a
function with each analyzer (including those staged in new_children).
This is a depth first traversal currently. The idea here is to have a
generic way to walk over the analyzer tree, rather than adding more
Analyzer member methods if all that is needed is to collect analyzers,
or call a certain member method on an analyzer.