Fix ssl certificate logging when encountering protocol errors
This change fixes a problem with ssl certificate logging when a protocol
error is encountered after certificates are sent over a SSL/TLS
connection, but before the connection is logged. Currently, in these
cases, `ssl.log` will not contain the fingerprints of the certifiates
contained in the connection. This is due to the necessary hook not being
called.

Fixes GH-5078