Allow VLAN of 0
Fixes #5105

A VLAN of 0 is valid, but Zeek was essentially ignoring them. This means
that the PCP/DEI bits wouldn't get set, and you'd have no way to tell
that it has a VLAN of 0. This makes it so that Zeek under the hood can
differentiate VLAN of 0 from the absence of a VLAN by using a too-high
value.

Note that this only changes script behavior slightly since any VLAN
fields were optional. The only change is now it gets sent ahead with the
VLAN of 0.