Move `Packet` VLAN fields to accessors
Since these are optional fields, it makes sense to use optionals for
them. We can't change the type with deprecation easily, so move accessing
those fields off to accessors. This also hides the `vlan_present` bool.
We can later move the VLAN fields to just hold optionals, if that seems
good, to ditch that `vlan_present` bool.